dead code removal | fixed <<#123 scrapy messes up logging>>
