+++ aoget/aogetapp.py
@@ -6,40 +6,10 @@ from view.main_window import MainWindow
 from util.qt_util import install_catch_all_exception_handler
 import logging
 from config.app_config import get_config_value, load_config_from_file, AppConfig, get_app_version
+from config.log_config import setup_logging
 from db.aogetdb import init_db
 
 logger = logging.getLogger(__name__)
-
-
-def setup_logging():
-    # Create a logger
-    logger = logging.getLogger()
-    level = (
-        logging.DEBUG if get_config_value(AppConfig.DEBUG) == "true" else logging.INFO
-    )
-    logger.setLevel(level)
-
-    # Create a file handler for output file
-    file_handler = logging.FileHandler(get_config_value(AppConfig.LOG_FILE_PATH))
-    file_handler.setLevel(level)
-    file_handler.setFormatter(
-        logging.Formatter(
-            '%(asctime)s :: %(levelname)s :: %(name)s:%(lineno)d :: %(message)s'
-        )
-    )
-
-    # Create a console handler for output to the console
-    console_handler = logging.StreamHandler()
-    console_handler.setLevel(level)
-    console_handler.setFormatter(
-        logging.Formatter(
-            '%(asctime)s :: %(levelname)s :: %(name)s:%(lineno)d :: %(message)s'
-        )
-    )
-
-    # Add both handlers to the logger
-    logger.addHandler(file_handler)
-    logger.addHandler(console_handler)
 
 
 def setup_config():
